Feat: getProfilePhoto 파일 분리
- charSelectModal에서도 사용
